refactor(storage): split schema.rs into migrations/ + lookup submodules
`primer-storage/src/schema.rs` had grown to 623 lines, holding four
distinct concerns in one file: the schema-version constant, the v1
baseline DDL, the seven-step additive migration chain (with each
version's DDL constants interleaved), and the lookup-table
validate-and-seed pass. It has been a directory module since the
`v4_tests` sibling landed, so the split is a natural extension.

Behaviour-preserving. New layout:

- `schema/mod.rs` (104) — `USER_VERSION` + `SCHEMA_SQL` (the v1
  baseline DDL) + a flat re-export façade, so every `crate::schema::<name>`
  path that `store/mod.rs`, `catalog.rs`, and `store/tests/` already use
  is unchanged. Declared visibilities are preserved exactly: `pub` for
  `SCHEMA_SQL` / `USER_VERSION` / `apply_v2_migrations` /
  `validate_and_seed_lookup`, `pub(crate)` for `apply_v3..v8_migrations`.
- `schema/migrations/{v2..v8}.rs` (32-102 each) — one file per version
  step. Each carries that version's DDL constants right next to the
  `apply_vN_migrations` function that runs them, per the existing repo
  rule that one-shot schema SQL reads best next to its logic. The
  section-banner comments (`// ─── v4 schema strings ───`) became `//!`
  module docs, so the substantive rationale (the JSON-in-TEXT note on
  v4, the cascade-design note on v5) is now the file's own header.
- `schema/migrations/mod.rs` (49) — chain doc, submodule declarations +
  re-exports, and the two shared introspection helpers
  (`column_exists` / `table_exists`) the guarded `ALTER`s use.
- `schema/lookup.rs` (80) — `validate_and_seed_lookup`.
- `schema/v4_tests.rs` — untouched. It reaches `SCHEMA_SQL`,
  `USER_VERSION`, `apply_v2..v7_migrations`, and
  `validate_and_seed_lookup` by bare name through `use super::*`, which
  now resolves against the `mod.rs` façade.

Adding v9 is now: add `migrations/v9.rs`, one `mod` + one `pub(crate) use`
line in `migrations/mod.rs`, call it from the open path, bump
`USER_VERSION`.

Verification:
- `cargo test -p primer-storage` — 154 passed / 0 failed, identical to
  the pre-split baseline.
- External pub surface byte-identical (4 symbols) and the `pub(crate)`
  surface byte-identical (6 symbols), both diffed against `main`.
- Order-insensitive line diff of all substantive (non-comment,
  non-import) lines old-vs-new is empty except for the five re-export
  façade lines deliberately added — no migration or DDL line lost,
  moved-but-altered, or duplicated.
- `cargo clippy --workspace --all-targets -- -D warnings` clean;
  `cargo fmt --all -- --check` clean; `cargo test --workspace` green
  (51 × `test result: ok`, 0 failed); `cargo doc -p primer-storage
  --document-private-items` clean under `RUSTDOCFLAGS=-D warnings`
  (the new intra-doc links resolve).

Docs updated for the new layout: the `primer-storage` and
"Schema migrations" bullets in CLAUDE.md, and every `schema.rs` link,
code-fence path, and the "Add a schema migration" recipe in
docs/devel/05-storage-and-sessions.md (12 stale links, all now
resolving). Two pre-existing doc nits fixed inline while there: the
recipe quoted CLAUDE.md's sentence as "Schema is at user_version 8."
(it reads "Schema is at version 8."), and CLAUDE.md called the
migration template "the existing v2..v7 chain" when the chain reaches
v8. README/ROADMAP need no change — this alters no capability, no
schema version, and no user-facing behaviour.
